Revert "Class survey"
