+++ .gitignore
@@ -4,8 +4,6 @@
 /node_modules
 /.pnp
 .pnp.js
-package-lock.json
-yarn.lock
 /.firebase
 
 # testing
