+++ package.json
@@ -3,12 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@blueprintjs/core": "^3.12.0",
-    "@blueprintjs/icons": "^3.5.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-responsive": "^6.1.1",
-    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3"
   },
   "scripts": {
