+++ src/App.css
@@ -1,3 +1,38 @@
-.container {
-  padding: 2.25rem 10% 0;
+.app {
+  text-align: center;
+  text-align: -webkit-center;
+  background-color: #282c34;
+  min-height: 100vh;
+}
+
+.logo {
+  height: 30vmin;
+}
+
+.header {
+  text-align: justify;
+  text-align-last: center;
+  -moz-text-align-last: center;
+  -webkit-text-align-last: center;
+  font-size: calc(10px + 1.5vmin);
+  color: white;
+  padding: 1.25rem 10%;
+}
+
+.generate-button {
+  background-color: #fa8334;
+  color: white;
+  display: inline-block;
+  font: inherit;
+  text-align: center;
+  white-space: nowrap;
+  vertical-align: middle;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+  border: 1px solid transparent;
+  padding: .375rem .75rem;
+  line-height: 1.5;
+  border-radius: .25rem;
 }
