+++ src/Stats.css
@@ -1,0 +1,12 @@
+.stats {
+  text-align: center;
+}
+.greeting {
+  color: #61dafb;
+}
+.stat-list {
+  padding-inline-start: 0px;
+}
+.stat {
+  list-style-type: none;
+}
