--- src/components/class-survey/LevelZero.css
@@ -1,19 +1,0 @@
-#level-zero {
-  color: #61dafb;
-}
-.class-img {
-  height: 50vmin;
-}
-.class-title {
-  float: right;
-}
-.level-area{
-  display: flex;
-  justify-content: space-between;
-}
-
-@media only screen and (max-width: 767px) {
-  .class-img {
-    height: 50vmin;
-  }
-}
