--- src/components/class-survey/index.js
@@ -1,4 +1,0 @@
-import LevelZero from './level-zero';
-import './LevelZero.css';
-
-export { LevelZero };
