--- src/components/common/index.js
@@ -1,3 +1,0 @@
-import ListItem from './list-item';
-
-export { ListItem };
