--- src/components/common/list-item.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-class ListItem extends Component {
-  render() {
-    const text = this.props.text;
-    return (
-      <li>{text}</li>
-    );
-  }
-}
-
-export default ListItem;
