--- src/components/footer/Footer.css
@@ -1,22 +1,0 @@
-.footer {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-}
-.button-div {
-  margin: .5rem;
-}
-#nav-button {
-  border-radius: 30px;
-  padding: 18px 25px;
-  outline: none;
-}
-
-@media only screen and (max-width: 767px) {
-  .footer {
-    justify-content: flex-start;
-  }
-}
