--- src/components/footer/index.js
@@ -1,4 +1,0 @@
-import Footer from './Footer';
-import './Footer.css';
-
-export { Footer };
