--- src/components/header/Header.css
@@ -1,6 +1,0 @@
-.navbar-logo {
-  height: 7vmin;
-}
-#nav-heading {
-  margin-right: 5px;
-}
