--- src/components/header/index.js
@@ -1,4 +1,0 @@
-import Header from './Header';
-import './Header.css';
-
-export { Header };
