--- src/components/stat-roller/index.js
@@ -1,5 +1,0 @@
-import AbilityScores from './AbilityScores';
-import Scores from './Scores';
-import './Scores.css';
-
-export { AbilityScores, Scores };
